refactor: extract ManageInstalledVersion component
